requirements.txt
ultralytics>=8.3.0
opencv-python>=4.9.0.80
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0.1
streamlit>=1.37.0
piexif>=1.1.3
